editor: split tools.ts into multiple files to avoid circular deps
