+++ front/src/applications/editor/tools/common.tsx
@@ -3,11 +3,8 @@ import { MapEvent } from 'react-map-gl';
 import { IconType } from 'react-icons/lib/esm/iconBase';
 import { TFunction } from 'i18next';
 
-import { EditorState } from '../../reducers/editor';
-import { Item, PositionnedItem } from '../../types';
-import { SelectZone } from './tools/SelectZone';
-import { CreateLine } from './tools/CreateLine';
-import { SelectItems } from './tools/SelectItems';
+import { EditorState } from '../../../reducers/editor';
+import { Item, PositionnedItem } from '../../../types';
 
 export interface CommonToolState {
   mousePosition: [number, number];
@@ -94,5 +91,3 @@ export interface Tool<S extends CommonToolState> {
     editorState: EditorState
   ) => JSX.Element | null | undefined;
 }
-
-export const Tools: Tool<any>[] = [SelectZone, CreateLine, SelectItems];
