+++ front/src/applications/editor/tools/index.tsx
@@ -1,0 +1,7 @@
+import { SelectZone } from './SelectZone';
+import { CreateLine } from './CreateLine';
+import { SelectItems } from './SelectItems';
+import { Tool } from './common';
+
+export const Tools: Tool<any>[] = [SelectZone, CreateLine, SelectItems];
+export * from './common';
